remove 3d error threshold.
Previously, the perpendicular distances of the reconstructed 3D point XYZ to the rays from the cameras were used as a cost to select 3D points. The threshold is removed now.

In the stereo linking, the reprojected error is used as cost function.

For the 3D cloud, no points were deleted because of the large error: 1/error is used as a weight for each point.