a N Rook Problem solver
The extension calculate following problem. Given a NxN binary matrix where one is only allow to place rook in the matix whose element is 1, emulating all possible configurations that. For each configuration, the number of rooks equals the rank of the matrix. (i.e. search for configuration with maximum number of rook).