add new function to merge temporally overlapped trajectories
The function `merge_trajectories` handles this situation.

If trajectory 1 range from frame 0 to frame 100, trajectory 2 extend from frame 98 to frame 200.

These trajectories should actually correspond to the same agent. Trajectory 1 was mistakenly linked to wrong positions in frame 99 and 100.

To merge the two, I calculate the position in frame 98 for both trajectories. If they are close, these two trajectories were replaced with a new trajectory, with the following formula.

`new_traj = traj_1[0 : 98] + raj_2[98 : 200]`